# src/AppForSEII2526.Web/OpenAPIs/swagger3.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForMovies.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Reviews/GetReviews": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "operationId": "GetReviews",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDetailDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Reviews/CreateReview": {
      "post": {
        "tags": [
          "Reviews"
        ],
        "operationId": "CreateReview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewForCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewForCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewForCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDetailDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
      "ReviewDetailDTO": {
        "required": [
          "country",
          "customerUserName",
          "driverType"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "customerUserName": {
            "maxLength": 50,
            "minLength": 3,
            "type": "string",
            "format": "email"
          },
          "country": {
            "maxLength": 10,
            "minLength": 3,
            "type": "string"
          },
          "driverType": {
            "minLength": 1,
            "type": "string"
          },
          "reviewItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewItemsDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReviewForCreateDTO": {
        "required": [
          "country",
          "customerUserName",
          "driverType"
        ],
        "type": "object",
        "properties": {
          "customerUserName": {
            "maxLength": 50,
            "minLength": 3,
            "type": "string",
            "format": "email"
          },
          "country": {
            "maxLength": 10,
            "minLength": 3,
            "type": "string"
          },
          "driverType": {
            "minLength": 1,
            "type": "string"
          },
          "reviewItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewItemsDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReviewItemsDTO": {
        "required": [
          "rating"
        ],
        "type": "object",
        "properties": {
          "carID": {
            "type": "integer",
            "format": "int32"
          },
          "model": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "color": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "fueltype": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "manufacturer": {
            "maxLength": 50,
            "minLength": 4,
            "type": "string",
            "nullable": true
          },
          "rating": {
            "maximum": 5,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "reviewDescription": {
            "maxLength": 500,
            "minLength": 5,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": {}
      }
    }
  }
}